requirements.txt
aniso8601==8.0.0
attrs==19.3.0
Flask==1.1.2
flask-restplus==0.13.0
Flask-Script==2.0.6
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
Resource==0.2.1
six==1.15.0
Werkzeug==0.16.0
zipp==3.1.0
mysql-connector-python==8.0.11
pyowm==3.2.0
flask-cors==3.0.10